+++ sickrage/core/queues/__init__.py
@@ -89,11 +89,8 @@ class srQueue(PriorityQueue):
 
         with self.lock:
             while not self.empty() and not self.stop.isSet():
-                # only start a new task if one isn't already going
-                if self.currentItem is None or not self.currentItem.inProgress():
-                    # if the thread is dead then the current item should be finished
+                if self.currentItem is None or not self.currentItem.inProgress:
                     if self.currentItem:
-                        self.currentItem.finish()
                         self.currentItem = None
 
                     if self.queue[0][0] >= self.min_priority: